cloud_functions/fetch_arxiv_daily/requirements.txt
requests==2.31.0
feedparser==6.0.10
google-cloud-firestore==2.10.4
pytz==2023.3